refactor(cart): move seedCartData into server.ts, extract overlay bridge, trim comments
- seedCartData lives in lib/cart/server.ts with the other server-only cart
  helpers; delete seed.ts.
- Extract the cart overlay bridge into components/cart/overlay-bridge.tsx
  instead of defining it inline in layout.tsx.
- Trim narrating comments; keep single-line hidden-constraint notes.
